refactor: address codec moved
